fix/ID-7-change-app-according-to-atomic-design
